Create main_v4.py
Version 4 with BeautifulSoup.

New features: Fetching of head and body of HTML. Currently all reviews ca be extracted. The format of date has yet to be changed and also the review id has to be yet extracted. Both are tag attributes. Also utf8 coding should be stated globally and not locally.